% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_fatal_cdc_tbi.R, R/add_tbi_icd10cm.R
\name{add_fatal_cdc_tbi}
\alias{add_fatal_cdc_tbi}
\alias{add_fatal_cdc_tbi}
\title{Add TBI as defined by CDC using ICD-10}
\usage{
add_fatal_cdc_tbi(data, diag_ecode_col)

add_fatal_cdc_tbi(data, diag_ecode_col)
}
\arguments{
\item{data}{the death dataset in which to add the new variables}

\item{underly_col}{The column index for the with the ICD-10 underlying cause of death}

\item{mult_col}{The column indices of the multiple-cause-of-death codes}

\item{data}{the death dataset in which to add the new variables}

\item{underly_col}{The column index for the with the ICD-10 underlying cause of death}

\item{mult_col}{The column indices of the multiple-cause-of-death codes}
}
\value{
The original dataset with the added fields. 1 means variable present and 0 otherwise

The original dataset with the added fields. 1 means variable present and 0 otherwise
}
\description{
cdc_inj: Any injury death, that is an underlying cause-of-death ICD-10 codes of V01–Y36, Y85–Y87, Y89, U01–U03.
cdc_tbi: TBI deaths are injury deaths (cdc_inj) that involved the following ICD-10 codes in the multiple-cause-of-death fields

add_tbi_10cm: TBI deaths are injury that involved the following ICD-10M codes in any diagnosis field
}
\details{
S01.0–S01.9 Open wound of head
S02.0, S02.1, S02.3, S02.7–S02.9 Fracture of skull and facial bones
S04.0 Injury of optic nerve and pathways
S06.0–S06.9 Intracranial injury
S07.0, S07.1, S07.8, S07.9 Crushing injury of head
S09.7–S09.9 Other and unspecified injuries of head
T01.0 Open wounds involving head with neck
T02.0 Fractures involving head with neck
T04.0 Crushing injuries involving head with neck
T06.0 Injuries of brain and cranial nerves with injuries
of nerves and spinal cord at neck level
T90.1, T90.2, T90.4, T90.5, T90.8, T90.9 Sequelae of injuries of head
}
\examples{
Examples will be added later

Examples will be added later

}
