% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ice_intent_mech.r
\name{add_ice_intent_mech}
\alias{add_ice_intent_mech}
\title{Add ICECI intent and mechanism of injury}
\usage{
add_ice_intent_mech(data, uid, underlying_cause, ignore_case = TRUE)
}
\arguments{
\item{data}{input data}

\item{uid}{unique id variable without quotation marks}

\item{underlying_cause}{index of the underlying cause variable}
}
\value{
return the input with additional variables (ice_intent, ice_mechanism)
}
\description{
, reference
}
\examples{

}
