% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_intent_mech_icd9cm.R
\name{add_intent_mech_icd9cm}
\alias{add_intent_mech_icd9cm}
\title{Add intent and mechanism combined fields for injury ICD-9-CM.}
\usage{
add_intent_mech_icd9cm(data, inj_col, reference = c("both", "intent",
  "mechanism"))
}
\arguments{
\item{data}{input data}

\item{inj_col}{ecode and diagnosis column indices}

\item{reference}{injury ice reference with three possible choices:
"both"(the default) adding intent and mechanism combined variables
"intent" adding intent only
"mechanism" adding mechanism only}
}
\value{
return the input with additional variables (intent_mechanism combinations, intent variables, or mechanism variables depending on the reference choice
}
\description{
, reference
}
\examples{
library(tidyverse)
dat <- data.frame(e1 = c("E9585", "E9885", "E986", "E966"),
e2 = c("E9050", "E9583", "E846", "E9556" ))

dat \%>\% create_intent_mech(inj_col = c(1,2), reference = "both")
dat \%>\% create_intent_mech(inj_col = c(1,2), reference = "intent")
dat \%>\% create_intent_mech(inj_col = c(1,2), reference = "mechanism")

}
