% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_self_harm_9_10cm.R
\name{add_self_harm_9_10cm}
\alias{add_self_harm_9_10cm}
\title{Finding Self-harm and suicide attempts}
\usage{
add_self_harm_9_10cm(data, diag_ecode_col, var_name, date)
}
\arguments{
\item{data}{the hospital dataset in which to add the new variables}

\item{diag_ecode_col}{The column indeces for the icd 9 and icd 10 cm}

\item{var_name}{The name for the new variable}

\item{date}{The name of the date of discharge field}
}
\value{
The original dataset with the added field. 1 means variable present and 0 otherwise
}
\description{
Finding Self-harm and suicide attempts
}
\examples{
Examples will be added later

}
