% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_to_year.r
\name{date_to_year}
\alias{date_to_year}
\title{Calculate number of years between two dates}
\usage{
date_to_year(date_start, date_end)
}
\arguments{
\item{date_start:}{start date}

\item{date_end:}{end date}
}
\value{
number of years
}
\description{
Calculate number of years between two dates
}
\examples{
to be added

}
