% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd9cm_intent_mech_regex}
\alias{icd9cm_intent_mech_regex}
\title{icd9 intent and mechanism.}
\format{Data frame}
\usage{
icd9cm_intent_mech_regex
}
\description{
Dataset of 84 rows and 4 variables.
}
\examples{

}
\keyword{reference}
