% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd9cm_mech_regex}
\alias{icd9cm_mech_regex}
\title{icd9 mechanism.}
\format{Data frame}
\usage{
icd9cm_mech_regex
}
\description{
Dataset of 30 rows and 3 variables.
}
\examples{

}
\keyword{reference}
