% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ice_code9}
\alias{ice_code9}
\title{ICECI codes for ICD10.}
\format{Data frame}
\usage{
ice_code9
}
\description{
Dataset of 1251 rows and 6 variables.
}
\examples{
library(dplyr)
sample_n(ice_code9, 5)

}
\keyword{reference}
