% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onclip.R
\name{onclip}
\alias{onclip}
\title{Copy to clipboard. A wrapper to write.table}
\usage{
onclip(x, row.names = F, col.names = T)
}
\arguments{
\item{x}{input data}

\item{row.names}{default to F}

\item{col.names}{default to Tl}
}
\value{
table saved to clipboard
}
\description{
, reference
}
\examples{

}
