% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbi_icd9cm.R
\name{tbi_icd9cm}
\alias{tbi_icd9cm}
\title{Find TBI based on ICD-9-CM.
800.00–801.99 Fracture of the vault or base of the skull
 803.00–804.99 Other and unqualified or multiple fractures of the skull
 850.0–850.9 Concussion
 851.00–854.19 Intracranial injury, including contusion, laceration, and hemorrhage
 950.1–950.3 Injury to the optic chiasm, optic pathways, or visual cortex
 959.01 Head injury, unspecified
 995.55 Shaken infant syndrome
Find any drug and other opioids from ICD-9-CM}
\usage{
tbi_icd9cm(data, diag_ecode_col)
}
\arguments{
\item{data:}{input data}

\item{diag_ecode_col:}{indices of diagnosis and E codes}
}
\value{
tbi_icd9cm
}
\description{
Find TBI based on ICD-9-CM.
800.00–801.99 Fracture of the vault or base of the skull
 803.00–804.99 Other and unqualified or multiple fractures of the skull
 850.0–850.9 Concussion
 851.00–854.19 Intracranial injury, including contusion, laceration, and hemorrhage
 950.1–950.3 Injury to the optic chiasm, optic pathways, or visual cortex
 959.01 Head injury, unspecified
 995.55 Shaken infant syndrome
Find any drug and other opioids from ICD-9-CM
}
\examples{
ibrary(tidyverse)
filter(hosp_set, year == 2014) \%>\% od_drug_types_icd9cm(diag_ecode_col = c(3, 6)) \%>\% sample_n(5)
}
